+++ considerations/front-react/package.json
@@ -10,6 +10,7 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "@xyflow/react": "^12.8.5",
     "react": "^19.1.1",
     "react-dom": "^19.1.1",
     "react-router": "^7.9.1"
