+++ considerations/front-react/vite.config.ts
@@ -2,6 +2,7 @@ import { defineConfig } from 'vite'
 import react from '@vitejs/plugin-react'
 import path from "node:path";
 
+// https://vite.dev/config/
 export default defineConfig({
   plugins: [react()],
   resolve: {
